requirements.txt
# Global Lib
python-dotenv==1.0.1
PyYAML==6.0.2
pytest==8.3.2
pytest-html==4.1.1
pytest-xdist==3.6.1
pytest-rerunfailures==12.0
pytest-asyncio==0.23.8

# Testing Lib
pytest-playwright>=0.5.0
playwright>=1.48.0
allure-pytest==2.13.5
httpx>=0.27.0

# Performance
locust==2.31.6

# Mobile
Appium-Python-Client==4.0.0

# Quality
ruff==0.5.7
black==24.8.0